Split shell LSP tests
